Move heuristic agents to package files
